requirements_production.txt
asgiref==3.7.2
click==8.1.6
diff-match-patch==20230430
Django==4.2.3
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-import-export==3.2.0
django-prometheus==2.3.1
gunicorn==20.1.0
Pillow==10.0.0
psycopg2-binary==2.9.6
qrcode==7.4.2
sqlparse==0.4.4
wait-for-it==2.2.1
